Refs 13763. Implementation of DDSFilterValue::compare.

+++ src/cpp/fastdds/topic/DDSSQLFilter/DDSFilterValue.hpp
@@ -156,7 +156,7 @@ protected:
 
 private:
 
-    static int compare(
+    static int64_t compare(
             const DDSFilterValue& lhs,
             const DDSFilterValue& rhs) noexcept;
